1570 Extract shared AnthropicClient and OpenAiApi beans into ai-config module
