+++ server/libs/config/eval-config/build.gradle.kts
@@ -1,0 +1,5 @@
+dependencies {
+    implementation("org.springframework:spring-context")
+    implementation(project(":server:libs:ai:ai-tool-api"))
+    implementation(project(":server:libs:core:evaluator:evaluator-impl"))
+}
